feat: Use sliver lists for better performance
